requirements.txt
asgiref==3.11.0
Django==6.0.1
django-filter==25.2
django-tables2==2.8.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
whitenoise==6.11.0
